Move login page in its own file
