+++ app/src/androidTest/java/com/github/se/stepquest/LoginScreenTest.kt
@@ -2,6 +2,9 @@ package com.github.se.stepquest
 
 import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.navigation.compose.rememberNavController
+import com.github.se.stepquest.screens.LoginScreen
+import com.github.se.stepquest.ui.navigation.NavigationActions
 import com.github.se.stepquest.ui.theme.StepQuestTheme
 import org.junit.Rule
 import org.junit.Test
@@ -13,7 +16,9 @@ class LoginScreenTest {
   @Test
   fun everythingIsDisplayed() {
 
-    composeTestRule.setContent { StepQuestTheme { AppNavigationHost() } }
+    composeTestRule.setContent {
+      StepQuestTheme { LoginScreen(NavigationActions(rememberNavController())) }
+    }
 
     // Application logo is displayed
     composeTestRule.onNodeWithTag("App logo").assertExists("The app logo doesn't exist?!")
@@ -26,11 +31,5 @@ class LoginScreenTest {
         .onNodeWithText("New player")
         .assertExists("The new player button doesn't exist?!")
     composeTestRule.onNodeWithText("New player").assertHasClickAction()
-  }
-
-  @Test
-  fun buttonsAreClickable() {
-
-    composeTestRule.setContent { StepQuestTheme { AppNavigationHost() } }
   }
 }
